requirements.txt
pandas
requests
google-cloud-bigquery
pyarrow
python-dotenv
pandera

protobuf